fix(security): sweep orphaned signatures in secure-docs too, not just project-files
The fail-closed retention in discardSignatureUnlessReferenced() and
persistOwnedSignature() deliberately keeps an uploaded e-signature when a
cleanup probe fails or is inconclusive, favoring a storage leak over risking
deletion of a committed signature. Those retained objects land in the
private secure-docs bucket, which the old public-only sweeper never scanned
-- they would accumulate indefinitely with nothing to clean them up.

- Rename sweep-orphaned-public-signatures.mjs -> sweep-orphaned-signature-objects.mjs
  (git mv) since it now also deletes from a private bucket; update every
  reference to the old filename (actions.ts comment, restore script comment,
  STORAGE-AUDIT.md runbook).
- Add --bucket <project-files|secure-docs|both> (default both) so a plain
  run covers everything. secure-docs is swept in full (it holds only
  in-scope sensitive documents by design); project-files keeps its existing
  signatures/ + sensitive-PDF-pattern scan.
- Reference-check logic is unchanged and already correct for secure: refs --
  a DB value like "secure:signatures/x/y.png" contains the bucket-relative
  path as a substring, so the existing full-path/basename match still works.
- Backups are now bucket-qualified (<dest>/<bucket>/<path>) so project-files
  and secure-docs objects at identical paths (secure-docs was populated by
  copying project-files paths 1:1) can never collide in the backup.
- Per-bucket + grand-total counters in the summary output.

Not run (touches production storage + DB) -- Justin will run it himself.
